+++ examples/google-service-account.py
@@ -37,8 +37,8 @@ url = "https://apidata.googleusercontent.com/caldav/v2/" + calid + "/events"
 
 client = get_davclient(url, auth=OAuth(creds))
 
-for calendar in client.principal().calendars():
-    events = calendar.events()
+for calendar in client.principal().get_calendars():
+    events = calendar.get_events()
     for event in events:
         ## Comment from caldav maintainer: this usage of vobject works out as
         ## long as there are only events (and no tasks) on the calendar and